Relocated ui states to a new state folder under ui for add product and my product view models.
