+++ src/commonMain/kotlin/onlytrade/app/viewmodel/product/ui/AddProductViewModel.kt
@@ -4,19 +4,20 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.AddProductFailed
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.CategoryNotSelected
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.DescriptionBlank
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.EstPriceBlank
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.EstPriceLow
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.Idle
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.ImagesNotSelected
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.LessImagesSelected
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.Loading
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.MoreImagesSelected
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.ProductInReview
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.SubcategoryNotSelected
-import onlytrade.app.viewmodel.product.ui.AddProductUIState.TitleBlank
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.AddProductFailed
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.CategoryNotSelected
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.DescriptionBlank
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.EstPriceBlank
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.EstPriceLow
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.Idle
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.ImagesNotSelected
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.LessImagesSelected
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.Loading
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.MoreImagesSelected
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.ProductInReview
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.SubcategoryNotSelected
+import onlytrade.app.viewmodel.product.ui.state.AddProductUIState.TitleBlank
 import onlytrade.app.viewmodel.product.ui.usecase.AddProductUseCase
 import onlytrade.app.viewmodel.product.ui.usecase.AddProductUseCase.Result
